fix: wrong memory address data type and skipping PC counter

+++ src/mbc/romonly.rs
@@ -14,10 +14,10 @@ impl RomOnly {
 }
 
 impl ReadMemory for RomOnly {
-    fn read_memory_u8(&self, address: usize) -> u8 {
+    fn read_memory_u8(&self, address: u16) -> u8 {
         match address {
             0x0000..=0x7FFF => {
-                *self.rom_data.get(address).unwrap()
+                *self.rom_data.get(address as usize).unwrap()
             },
             0xA000..=0xBFFF => {
                 todo!("external ram")
@@ -26,17 +26,17 @@ impl ReadMemory for RomOnly {
         }
     }
 
-    fn read_memory_u16(&self, address: usize) -> u16 {
+    fn read_memory_u16(&self, address: u16) -> u16 {
         todo!()
     }
 }
 
 impl WriteMemory for RomOnly {
-    fn write_memory_u8(&mut self, address: usize, value: u8) {
+    fn write_memory_u8(&mut self, address: u16, value: u8) {
         todo!()
     }
 
-    fn write_memory_u16(&mut self, address: usize, value: u16) {
+    fn write_memory_u16(&mut self, address: u16, value: u16) {
         todo!()
     }
 }
